feat: Have robot follow shortest path to goal
If pathFollow flag is true the robot will follow the shortest calculated path to the goal, else the robot will apply a bug algorithm to find te goal. It will move straight to the goal until an obstacle is encountered and will then do obstacle avoidance around obstalse. Once the obstacle is cleared the robot will once again try to reach the goal directly