feat: add persistant gravity to kinect data within a specific viewign distance from the kinect
Persistant gravity is used to ensure obstacles seen by the Kinect is preserved after it is outside the FOW of the kinect.
The idea is to remember obstacles so that these obstacles can fill in the dead zone of the kinect. This will help with obstacle avoidance and path planning